refactor(deepseek_v4): move Sinkhorn to shared module
Moved _make_sinkhorn_kernel and hc_split_sinkhorn from deepseek_v4.py
to mlx_lm/models/sinkhorn.py as a shared module, per pcuenca's review
comment. Other models using doubly-stochastic mixing can now import
the fused Metal kernel directly.

No functional changes. 8/8 tests pass, benchmark unchanged.
